requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
typing_extensions==4.15.0

# REST API
djangorestframework==3.15.2

# PostgreSQL driver
psycopg[binary]==3.2.3

# Déploiement Render
gunicorn==23.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
dj-database-url==2.2.0
django-environ==0.11.2
